# api-gateway/api/docs/swagger.yaml
basePath: /v1
definitions:
  chat_service.UserReq:
    properties:
      bio:
        type: string
      first_name:
        type: string
      image:
        type: string
      last_name:
        type: string
      phone:
        type: string
      user_name:
        type: string
    type: object
  chat_service.UserRes:
    properties:
      bio:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone:
        type: string
      user_name:
        type: string
    type: object
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: example@gmail.com
    name: Murtazoxon Gofurov
  description: This is a sample server celler server.
  title: ChatApp API
  version: "1.0"
paths:
  /user:
    post:
      consumes:
      - application/json
      description: post string user info
      parameters:
      - description: Post user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/chat_service.UserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat_service.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - BearerAuth: []
      summary: Add an user info
      tags:
      - User
  /user/{id}:
    get:
      consumes:
      - application/json
      description: get string user info
      parameters:
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat_service.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
      summary: Get an user info
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"